+++ src/main/java/io/fabric8/maven/docker/access/chunked/EntityStreamReaderUtil.java
@@ -12,20 +12,27 @@ public class EntityStreamReaderUtil {
     private EntityStreamReaderUtil() {}
 
     public static void processJsonStream(JsonEntityResponseHandler handler, InputStream stream) throws IOException {
-        JSONTokener tokener = new JSONTokener(stream);
-        while (true) {
-            char next = tokener.nextClean();
-            if (next == 0) {
-                return;
-            } else {
-                tokener.back();
+        handler.start();
+        try {
+            JSONTokener tokener = new JSONTokener(stream);
+            while (true) {
+                char next = tokener.nextClean();
+                if (next == 0) {
+                    return;
+                } else {
+                    tokener.back();
+                }
+                JSONObject object = new JSONObject(tokener);
+                handler.process(object);
             }
-            JSONObject object = new JSONObject(tokener);
-            handler.process(object);
+        } finally {
+            handler.stop();
         }
     }
 
     public interface JsonEntityResponseHandler {
         void process(JSONObject toProcess) throws DockerAccessException;
+        void start();
+        void stop();
     }
 }