+++ src/main/java/io/fabric8/maven/docker/access/hc/DockerAccessWithHcClient.java
@@ -444,5 +444,4 @@ public class DockerAccessWithHcClient implements DockerAccess {
             return null;
         }
     }
-
 }