+++ doc/changelog.md
@@ -1,5 +1,8 @@
 # ChangeLog
 
+* **0.15.1-SNAPSHOT
+  - Fix push / pull progress bar (#91)
+
 * **0.15.0** (2016-04-27)
   - Be more conservative when no "warnings" are returned on create ([#407](https://github.com/fabric8io/docker-maven-plugin/issues/407))
   - Fix parsing of timestamps with numeric timezone ([#410](https://github.com/fabric8io/docker-maven-plugin/issues/410))